Tidy up performance tracing
